feat: extract reusable component system
Create src/lib/components/ with five components:
- Alert: error/success alert box
- PageContainer: max-width wrapper with padding
- ScoreSubmitForm: submit score section with inline alerts
- Leaderboard: full leaderboard with game/time filter state
- RecentSubmissions: recent submissions table

Reduces groups/[id]/+page.svelte from 230 → 63 lines.
Both groups pages now use PageContainer and Alert.
